pyproject.toml
[project]
name = "lcd-remote-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "jinja2",
    "uvicorn",
    "luma.lcd",
    "pillow",
    "rpi-gpio",
    "spidev",
    "numpy",
    "gpiozero",
]
